+++ packages/react-widgets/src/hooks/useWidgetFetch.js
@@ -1,7 +1,8 @@
 import {
   InvalidColumnError,
   getGeometryToIntersect,
-  isGlobalViewport
+  isGlobalViewport,
+  normalizeGeometry
 } from '@carto/react-core';
 import {
   selectAreFeaturesReadyForSource,
@@ -15,6 +16,40 @@ import { DEFAULT_INVALID_COLUMN_ERR } from '../widgets/utils/constants';
 import useCustomCompareEffect from './useCustomCompareEffect';
 import useWidgetSource from './useWidgetSource';
 import { isRemoteCalculationSupported } from '../models/utils';
+
+/**
+ * Select a geometry to intersect features with, given the widget configuration.
+ *
+ * @typedef { import('geojson').Polygon | import('geojson').MultiPolygon } Geometry
+ * @typedef { import('geojson').Feature<Geometry> } Feature
+ * @typedef { import('@carto/react-core').Viewport } Viewport
+ *
+ * @param {boolean} global if the widget is in global mode
+ * @param {Viewport?} viewport the current viewport
+ * @param {Feature?} spatialFilter the current active mask, if set
+ * @returns {Geometry?}
+ *   the geometry to intersect, if any, to be used as spatial filter for queries
+ */
+export function selectGeometryToIntersect(global, viewport, spatialFilter) {
+  if (global) {
+    // global widgets work with selection mask, if set, otherwise no filter at all
+    if (spatialFilter?.geometry?.coordinates) {
+      return normalizeGeometry(spatialFilter.geometry);
+    } else {
+      return null;
+    }
+  } else {
+    // shortcut to remove viewports that cover the world anyway
+    if (!spatialFilter && isGlobalViewport(viewport)) {
+      return null;
+    } else {
+      // non-global widgets work with selection mask, if set, otherwise viewport
+      return normalizeGeometry(
+        getGeometryToIntersect(viewport, spatialFilter ? spatialFilter.geometry : null)
+      );
+    }
+  }
+}
 
 export default function useWidgetFetch(
   modelFn,
@@ -44,10 +79,7 @@ export default function useWidgetFetch(
 
   const viewport = useSelector(selectViewport);
   const spatialFilter = useSelector((state) => selectSpatialFilter(state, dataSource));
-  const geometryToIntersect =
-    global || (!spatialFilter && isGlobalViewport(viewport))
-      ? null
-      : getGeometryToIntersect(viewport, spatialFilter ? spatialFilter.geometry : null);
+  const geometryToIntersect = selectGeometryToIntersect(global, viewport, spatialFilter);
 
   useCustomCompareEffect(
     () => {