+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Supporting for `client` parameter for Widgets API calls without passing client as component attribute [#729](https://github.com/CartoDB/carto-react/pull/729)
 - Calculation of widget using maps API under FF [#658](https://github.com/CartoDB/carto-react/pull/658)
+- Mask, when set, is applied to global widgets as well as to viewport-based widgets [#704](https://github.com/CartoDB/carto-react/pull/704)
+- Support for remote scatter plot widget [#704](https://github.com/CartoDB/carto-react/pull/704)
 
 ## 1.5